requirements.txt
astroid==2.7.3
autopep8==1.5.7
click==8.0.1
Flask==2.0.1
importlib-metadata==4.8.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.0.26
platformdirs==2.3.0
protobuf==3.17.3
pycodestyle==2.7.0
pylint==2.10.2
six==1.16.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
